requirements.txt
blinker==1.9.0
click==8.3.1
Flask==3.1.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysql-connector-python==9.6.0
python-dotenv==1.2.1
Werkzeug==3.1.6
